fix: applying suggestions
